rename downsample file for consistency
