+++ lib/sdwnscalculator.js
